test: split admin basic endpoint tests
