+++ crates/dcc-mcp-gateway/src/gateway/admin/tests.rs
@@ -425,28 +425,6 @@ filters:
             .status()
     }
 
-    async fn body_html(router: Router, uri: &str) -> (StatusCode, String, String) {
-        let resp = router
-            .oneshot(
-                Request::builder()
-                    .uri(uri)
-                    .body(axum::body::Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        let status = resp.status();
-        let ct = resp
-            .headers()
-            .get("content-type")
-            .and_then(|v| v.to_str().ok())
-            .unwrap_or("")
-            .to_string();
-        let bytes = to_bytes(resp.into_body(), 4 * 1024 * 1024).await.unwrap();
-        let body = String::from_utf8_lossy(&bytes).to_string();
-        (status, ct, body)
-    }
-
     // ── HTML dashboard ────────────────────────────────────────────────────
 
     #[tokio::test]
@@ -481,147 +459,6 @@ filters:
         assert!(doc["paths"].get("/v1/debug/analytics/export").is_some());
         assert!(doc["paths"].get("/v1/debug/deregistered").is_some());
         assert!(doc["paths"].get("/v1/debug/integrations").is_some());
-    }
-
-    #[tokio::test]
-    async fn test_admin_ui_returns_html() {
-        let (status, ct, _) = body_html(admin_router(), "/").await;
-        assert_eq!(status, StatusCode::OK);
-        assert!(ct.contains("text/html"), "expected text/html, got {ct}");
-    }
-
-    #[tokio::test]
-    async fn test_admin_html_has_title() {
-        let (_, _, html) = body_html(admin_router(), "/").await;
-        assert!(
-            html.contains("<title>") && (html.contains("DCC-MCP") || html.contains("Admin")),
-            "HTML missing expected <title> content"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_html_contains_api_references() {
-        let (_, _, html) = body_html(admin_router(), "/").await;
-        for endpoint in &["instances", "tools", "health", "traces", "stats"] {
-            assert!(
-                html.contains(endpoint),
-                "HTML missing reference to '{endpoint}'"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_admin_html_contains_traces_and_stats_panels() {
-        let (_, _, html) = body_html(admin_router(), "/").await;
-        // Vite minifies JSX; assert stable API paths and panel strings from the bundle.
-        for needle in [
-            "/traces?limit=",
-            "/stats?range=",
-            "trace-row",
-            "No traces recorded.",
-        ] {
-            assert!(html.contains(needle), "HTML missing {needle}");
-        }
-        assert!(
-            html.contains("data-panel"),
-            "HTML missing data-panel attribute hooks"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_html_is_valid_doctype() {
-        let (_, _, html) = body_html(admin_router(), "/").await;
-        let trimmed = html.trim_start().to_lowercase();
-        assert!(
-            trimmed.starts_with("<!doctype html>"),
-            "HTML must start with <!DOCTYPE html>"
-        );
-    }
-
-    // ── /api/instances ────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_admin_instances_returns_json_array() {
-        let (status, body) = body_json(admin_router(), "/api/instances").await;
-        assert_eq!(status, StatusCode::OK);
-        assert!(
-            body["instances"].is_array(),
-            "expected 'instances' array, got {body}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_instances_empty_without_dccs() {
-        let (_, body) = body_json(admin_router(), "/api/instances").await;
-        assert!(body["instances"].as_array().unwrap().is_empty());
-    }
-
-    // ── /api/health ───────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_admin_health_returns_ok() {
-        let (status, body) = body_json(admin_router(), "/api/health").await;
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(
-            body["status"].as_str(),
-            Some("ok"),
-            "expected status=ok, got {body}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_health_has_uptime_secs() {
-        let (_, body) = body_json(admin_router(), "/api/health").await;
-        assert!(
-            body["uptime_secs"].as_u64().is_some(),
-            "expected uptime_secs >= 0"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_health_instances_total_is_zero() {
-        let (_, body) = body_json(admin_router(), "/api/health").await;
-        assert_eq!(body["instances_total"].as_u64(), Some(0));
-    }
-
-    #[tokio::test]
-    async fn test_admin_health_has_instances_ready_field() {
-        let (_, body) = body_json(admin_router(), "/api/health").await;
-        assert!(
-            body.get("instances_ready").is_some(),
-            "expected instances_ready field"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_health_includes_limits_and_circuits() {
-        let (_, body) = body_json(admin_router(), "/api/health").await;
-        assert!(body.get("limits").is_some(), "expected limits object");
-        assert!(body.get("circuits").is_some(), "expected circuits object");
-        assert!(body.get("rss_bytes").is_some(), "expected rss_bytes field");
-        assert_eq!(body["response_format"]["default"], "toon");
-        assert_eq!(
-            body["response_format"]["token_estimator"],
-            "dcc-mcp-byte4-v1"
-        );
-    }
-
-    // ── /api/tools ────────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_admin_tools_returns_json_array() {
-        let (status, body) = body_json(admin_router(), "/api/tools").await;
-        assert_eq!(status, StatusCode::OK);
-        assert!(
-            body["tools"].is_array(),
-            "expected 'tools' array, got {body}"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_admin_tools_empty_without_dccs() {
-        let (_, body) = body_json(admin_router(), "/api/tools").await;
-        assert!(body["tools"].as_array().unwrap().is_empty());
     }
 
     #[tokio::test]
@@ -1663,59 +1500,6 @@ filters:
         assert_eq!(limited_status, StatusCode::OK);
         assert_eq!(limited_body["logs"].as_array().unwrap().len(), 1);
         assert_eq!(limited_body["total"], 1);
-    }
-
-    // ── unknown routes ────────────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_admin_unknown_path_returns_404() {
-        let resp = admin_router()
-            .oneshot(
-                Request::builder()
-                    .uri("/api/doesnotexist")
-                    .body(axum::body::Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(resp.status(), StatusCode::NOT_FOUND);
-    }
-
-    // ── content-type headers ──────────────────────────────────────────────
-
-    #[tokio::test]
-    async fn test_json_endpoints_content_type() {
-        for uri in [
-            "/api/instances",
-            "/api/health",
-            "/api/tools",
-            "/api/skills",
-            "/api/calls",
-            "/api/logs",
-            "/api/stats",
-            "/api/governance",
-            "/api/traces",
-            "/api/workflows",
-        ] {
-            let resp = admin_router()
-                .oneshot(
-                    Request::builder()
-                        .uri(uri)
-                        .body(axum::body::Body::empty())
-                        .unwrap(),
-                )
-                .await
-                .unwrap();
-            let ct = resp
-                .headers()
-                .get("content-type")
-                .and_then(|v| v.to_str().ok())
-                .unwrap_or("");
-            assert!(
-                ct.contains("application/json"),
-                "endpoint {uri} must return application/json, got '{ct}'"
-            );
-        }
     }
 
     // ── /api/workers (Phase 4) ────────────────────────────────────────────
